% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CME.R
\name{spacejam}
\alias{spacejam}
\title{API connection}
\usage{
spacejam(start_date, end_date)
}
\arguments{
\item{start_date}{character}

\item{end_date}{character}
}
\value{
dataframe
}
\description{
API connection
}
